audio-to-text-vs-cloud/requirements.txt
openai-whisper
pandas
matplotlib
tqdm
requests
soundfile
tabulate